requirements.txt
streamlit
pandas
plotly
numpy
matplotlib
seaborn
google-cloud-storage